server/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
python-dotenv
pypdf
